Working with files and images

+++ SHOP_Mongoose/controllers/admin.js
@@ -1,7 +1,10 @@
 
 const mongodb = require('mongodb');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
+
+const { validationResult } = require('express-validator');
 
 
 exports.getAddProduct = (req, res, next) => {
@@ -11,8 +14,10 @@ exports.getAddProduct = (req, res, next) => {
     res.render('admin/editProduct', {
     pageTitle: 'Add Product', 
     path: '/admin/addProduct', 
-    editing: false
-
+    editing: false,
+    hasError: false,
+    errorMessage: null,
+    validationErrors: []
     });
 };
 
@@ -21,14 +26,48 @@ exports.getAddProduct = (req, res, next) => {
 exports.postAddProduct = (req, res, next) => {
     
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image){
+        return res.status(422).render('admin/editProduct', {
+            pageTitle: 'Add Product', 
+            path: '/admin/editProduct',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        }); 
+    }
+    const errors = validationResult(req);
+
+    if(!errors.isEmpty()){
+        console.log(errors.array());
+        return res.status(422).render('admin/editProduct', {
+            pageTitle: 'Add Product', 
+            path: '/admin/editProduct',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: errors.array()[0].msg,
+            validationErrors: errors.array()
+        });
+    }
+
     const product = new Product({
         title: title, 
         price: price, 
         description: description, 
-        imageUrl: imageUrl,
+        imageUrl: image.path,
         userId: req.user
     });
     product.save()
@@ -36,7 +75,11 @@ exports.postAddProduct = (req, res, next) => {
         console.log('Created Product Through Mongoose'); 
         res.redirect('/admin/products');
     })
-    .catch(err => { console.log(err) });
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
 
 
@@ -55,18 +98,44 @@ exports.getEditProduct = (req, res, next) => {
             pageTitle: 'Edit Product', 
             path: '/admin/editProduct',
             editing: editMode,
-            product: product
+            product: product,
+            hasError: false,
+            errorMessage: null,
+            validationErrors: []
         });
     })
-    .catch(err => { console.log(err) });
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const updatedPrice = req.body.price;
+    const errors = validationResult(req);
+
+    if(!errors.isEmpty()){
+        return res.status(422).render('admin/editProduct', {
+            pageTitle: 'Edit Product', 
+            path: '/admin/editProduct',
+            editing: true,
+            hasError: true,
+            product: {
+                title: updatedTitle,
+                price: updatedPrice,
+                description: updatedDescription,
+                _id: prodId
+            },
+            errorMessage: errors.array()[0].msg,
+            validationErrors: errors.array()
+        });
+    }
+
     Product.findById(prodId).then(product => {
         if (product.userId.toString() !== req.user._id.toString()){
             return res.redirect('/');
@@ -74,13 +143,20 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice; 
         product.description = updatedDescription; 
-        product.imageUrl = updatedImageUrl;
+        if(image){
+            fileHelper.deleteFiles(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save().then(result => { 
             console.log("Updated Product"); 
             res.redirect('/admin/products'); 
         })
     })
-    .catch(err => { console.log(err) });
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
 
 
@@ -96,7 +172,11 @@ exports.getProducts = (req, res, next) => {
             path: '/admin/products'
         });
     })
-    .catch(err => { console.log(err) });
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
 
 
@@ -109,10 +189,21 @@ If we want the deleted document then we can use findByIdAndRemove otherwise can 
 Recommend:- If don't want to get the deleted document then have to use findByIdAndDelete because it's fast cause does not return the document. 
 */
 exports.postDeleteProduct = (req, res, next) => {
-   const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-    res.redirect('/admin/products')
+    const prodId = req.body.productId;
+    Product.findById(prodId).then(prod => {
+        if(!prod)
+        {
+            return next(new Error('Product Not Found'));
+        }
+        fileHelper.deleteFiles(prod.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id })
     })
-   .catch(err => { console.log(err) });
+    .then(() => {
+        res.redirect('/admin/products')
+    })
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 };
